refactor: изменить структуру папок проекта
