+++ src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -3,7 +3,7 @@ package ru.practicum.shareit.item.model;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import ru.practicum.shareit.user.User;
+import ru.practicum.shareit.user.model.User;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
